chore(desktop): move export transactions from common to desktop
