+++ packages/suite/src/views/wallet/transactions/TransactionList/TransactionListActions/ExportAction.tsx
@@ -6,7 +6,8 @@ import { useDispatch } from 'src/hooks/suite';
 import { useTranslation } from 'src/hooks/suite/useTranslation';
 import { useSelector } from 'src/hooks/suite/useSelector';
 import { notificationsActions } from '@suite-common/toast-notifications';
-import { exportTransactionsThunk, fetchTransactionsThunk } from '@suite-common/wallet-core';
+import { fetchTransactionsThunk } from '@suite-common/wallet-core';
+import { exportTransactionsThunk } from 'src/actions/wallet/exportTransactionsActions';
 import { ExportFileType } from '@suite-common/wallet-types';
 import { Account } from 'src/types/wallet';
 import { isFeatureFlagEnabled, getTxsPerPage } from '@suite-common/suite-utils';
