+++ suite-common/wallet-utils/src/index.ts
@@ -4,7 +4,6 @@ export * from './balanceUtils';
 export * from './cardanoUtils';
 export * from './csvParserUtils';
 export * from './ethUtils';
-export * from './exportTransactionsUtils';
 export * from './fiatConverterUtils';
 export * from './fiatRatesUtils';
 export * from './formDraftUtils';
